Move cloneHelpers from products/utils to shared/services
- Fix backwards dependency: shared/services/DHTUploadService importing from products
- Move cloneHelpers.ts to shared/services/ where it belongs (infrastructure code)
- Update all imports: DHTUploadService, SearchApiClient, IndexGenerationService, SimpleCloneCache
- Maintains same functionality, improves architecture
