+++ ui/src/products/utils/SimpleCloneCache.ts
@@ -3,7 +3,7 @@
  */
 
 import type { AppClient, CellId } from "@holochain/client";
-import { getActiveCloneCellId } from "./cloneHelpers";
+import { getActiveCloneCellId } from "../../shared/services/cloneHelpers";
 import { startCloneSetup, updateCloneSetup, finishCloneSetup } from "../../shared/stores/LoadingStore";
 
 export class SimpleCloneCache {
